Implemented lazy create published repository
