+++ src/main/java/org/craftercms/studio/impl/v1/service/site/SiteServiceImpl.java
@@ -463,15 +463,7 @@ public class SiteServiceImpl implements SiteService {
                 logger.info("Syncing database with repository.");
                 syncDatabaseWithRepo(siteId, contentRepository.getRepoFirstCommitId(siteId), true);
 
-                // initial deployment
-                logger.info("Performing initial deployment");
-                contentRepository.initialPublish(siteId, sandboxBranch, servicesConfig.getLiveEnvironment(siteId),
-                        securityService.getCurrentUser(), "Create site.");
-                if (servicesConfig.isStagingEnvironmentEnabled(siteId)) {
-                    contentRepository.initialPublish(siteId, sandboxBranch, servicesConfig.getStagingEnvironment(siteId),
-                            securityService.getCurrentUser(), "Create site.");
-                }
-                objectStateService.setStateForSiteContent(siteId, State.EXISTING_UNEDITED_UNLOCKED);
+                objectStateService.setStateForSiteContent(siteId, State.NEW_UNPUBLISHED_UNLOCKED);
 	        } catch(Exception e) {
 	            success = false;
 	            logger.error("Error while creating site: " + siteName + " ID: " + siteId + " from blueprint: " +
@@ -718,7 +710,7 @@ public class SiteServiceImpl implements SiteService {
             // create site by cloning remote git repo
             logger.info("Creating site " + siteId + " by cloning remote repository " + remoteName +
                 " (" + remoteUrl + ")");
-            success = contentRepository.createSiteCloneRemote(siteId, sandboxBranch, remoteName, remoteUrl,
+            success = contentRepositoryV2.createSiteCloneRemote(siteId, sandboxBranch, remoteName, remoteUrl,
                 remoteBranch, singleBranch, authenticationType, remoteUsername, remotePassword, remoteToken,
                 remotePrivateKey, params, createAsOrphan);
 
@@ -763,7 +755,7 @@ public class SiteServiceImpl implements SiteService {
                 logger.error("Error while creating site: " + siteId + " ID: " + siteId + " as clone from" +
                              " remote repository: " + remoteName + " (" + remoteUrl + "). Rolling back...", e);
 
-                contentRepository.removeRemote(siteId, remoteName);
+                contentRepositoryV2.removeRemote(siteId, remoteName);
                 boolean deleted = contentRepository.deleteSite(siteId);
 
                 if (!deleted) {
@@ -809,15 +801,7 @@ public class SiteServiceImpl implements SiteService {
                 logger.info("Sync database with repository for site " + siteId);
                 syncDatabaseWithRepo(siteId, contentRepository.getRepoFirstCommitId(siteId), true);
 
-                // initial deployment
-                logger.info("Executing initial deployement for site " + siteId);
-                contentRepository.initialPublish(siteId, sandboxBranch, servicesConfig.getLiveEnvironment(siteId),
-                        securityService.getCurrentUser(), "Create site.");
-                if (servicesConfig.isStagingEnvironmentEnabled(siteId)) {
-                    contentRepository.initialPublish(siteId, sandboxBranch, servicesConfig.getStagingEnvironment(siteId),
-                            securityService.getCurrentUser(), "Create site.");
-                }
-                objectStateService.setStateForSiteContent(siteId, State.EXISTING_UNEDITED_UNLOCKED);
+                objectStateService.setStateForSiteContent(siteId, State.NEW_UNPUBLISHED_UNLOCKED);
             } catch(Exception e) {
                 success = false;
                 logger.error("Error while creating site: " + siteId + " ID: " + siteId + " as clone from " +
@@ -949,7 +933,7 @@ public class SiteServiceImpl implements SiteService {
                         InvalidRemoteUrlException | ServiceLayerException e) {
                     logger.error("Error while pushing site: " + siteId + " ID: " + siteId + " to remote repository "
                             + remoteName + " (" + remoteUrl + ")", e);
-                    contentRepository.removeRemote(siteId, remoteName);
+                    contentRepositoryV2.removeRemote(siteId, remoteName);
                 }
 
             try {
@@ -968,15 +952,7 @@ public class SiteServiceImpl implements SiteService {
                     logger.info("Sync database with repository for site " + siteId);
                     syncDatabaseWithRepo(siteId, contentRepository.getRepoFirstCommitId(siteId), true);
 
-                    // initial deployment
-                    logger.info("Executing initial deployement for site " + siteId);
-                    contentRepository.initialPublish(siteId, sandboxBranch, servicesConfig.getLiveEnvironment(siteId),
-                            securityService.getCurrentUser(), "Create site.");
-                    if (servicesConfig.isStagingEnvironmentEnabled(siteId)) {
-                        contentRepository.initialPublish(siteId, sandboxBranch, servicesConfig.getStagingEnvironment(siteId),
-                                securityService.getCurrentUser(), "Create site.");
-                    }
-                    objectStateService.setStateForSiteContent(siteId, State.EXISTING_UNEDITED_UNLOCKED);
+                    objectStateService.setStateForSiteContent(siteId, State.NEW_UNPUBLISHED_UNLOCKED);
                 } catch (Exception e) {
                     success = false;
                     logger.error("Error while creating site: " + siteId + " ID: " + siteId + " from blueprint: " +
@@ -1769,7 +1745,7 @@ public class SiteServiceImpl implements SiteService {
         if (!exists(siteId)) {
             throw new SiteNotFoundException();
         }
-        boolean toRet = contentRepository.removeRemote(siteId, remoteName);
+        boolean toRet = contentRepositoryV2.removeRemote(siteId, remoteName);
         insertRemoveRemoteAuditLog(siteId, remoteName);
         return toRet;
     }
